// src/app/core/models/create-policy-holder.ts
export interface CreatePolicyHolder {

    firstName: string;

    lastName: string;

    dateOfBirth: string;

    gender: string;

    email: string;

    phoneNumber: string;

    addressLine1: string;

    addressLine2?: string;

    city: string;

    state: string;

    postalCode: string;

    country: string;

}